refactor(utils): standalone image manipulation mod
utils has a standalone module for image manipulation. fixed improts in
all other modules
