+++ k2_oai/utils/__init__.py
@@ -3,5 +3,4 @@ Various utilities for the K2 OAI project. See the README.md for more information
 """
 
 from ._args_checker import is_positive_odd_integer, is_valid_method
-from ._image_manipulation import *
 from ._parsers import parse_str_as_coordinates
